rhsecapi: Remove the "less" pager use
There seems to be a problem within argparse for 3.7 which results in an
error around file handling. The process of resolving it would be too
arduous when compared to the alternative of avoiding calling "less"
directly in the manner previously used.

The recommended alternative is to simply pipe the help output to "less".